Working version of NAR packaged Pty4J (on Linux only so far). This will build the native parts, and the libraries will be automatically extracted and loaded on any clients of pty4j.
Although NAR plugin provides 'NarSystem' to do the initialisation, this is JNI focussed, and we need to use JNA to map to the interface. This is handles by NARPtyExecutor that was based on NarSystem, but only locates the library, and looks for shared libraries instead of JNI ones.

Clients of Pty4J should include both the main Jar and the native portions.

        <dependency>
			<groupId>pty4j</groupId>
			<artifactId>pty4j</artifactId>
			<version>0.5-SSHTOOLS-SNAPSHOT</version>
			<type>jar</type>
			<scope>compile</scope>
		</dependency>
		<dependency>
			<groupId>pty4j</groupId>
			<artifactId>pty4j</artifactId>
			<classifier>amd64-Linux-gpp-shared</classifier>
			<version>0.5-SSHTOOLS-SNAPSHOT</version>
			<type>nar</type>
		</dependency>